run: Only build binary package if not already built

+++ run.sh
@@ -68,18 +68,22 @@ if [ ! -f "${builddir}/${package_name}_${deb_version}.dsc" ]; then
     rm -r "${builddir}/${package_full}"
 fi
 
-run_prep || true
+# Generate binary package for this arch if not found
+build_arch=$(dpkg --print-architecture)
+if [ ! -f "${builddir}/${package_name}_${deb_version}_${build_arch}.buildinfo" ]; then
+    run_prep || true
 
-# Extract source package
-if [ ! -d "${builddir}/${package_name}_${deb_version}" ]; then
-    dpkg-source -x "${builddir}/${package_name}_${deb_version}.dsc" "${builddir}/${package_name}_${deb_version}"
+    # Extract source package
+    if [ ! -d "${builddir}/${package_name}_${deb_version}" ]; then
+        dpkg-source -x "${builddir}/${package_name}_${deb_version}.dsc" "${builddir}/${package_name}_${deb_version}"
+    fi
+
+    # Install build dependencies
+    (cd "${builddir}/${package_name}_${deb_version}" && mk-build-deps -ir -t "apt-get -o Debug::pkgProblemResolver=yes -y --no-install-recommends")
+
+    # Build binary package
+    (cd "${builddir}/${package_name}_${deb_version}" && debuild --no-lintian --no-sign)
+
+    # Cleanup intermediate build directory
+    rm -r "${builddir}/${package_name}_${deb_version}"
 fi
-
-# Install build dependencies
-(cd "${builddir}/${package_name}_${deb_version}" && mk-build-deps -ir -t "apt-get -o Debug::pkgProblemResolver=yes -y --no-install-recommends")
-
-# Build binary package
-(cd "${builddir}/${package_name}_${deb_version}" && debuild --no-lintian --no-sign)
-
-# Cleanup intermediate build directory
-rm -r "${builddir}/${package_name}_${deb_version}"
